sunset google.golang.org/appengine module
The appengine packages that were designed for the go1 runtime, were left
functional in go111, but Google is requesting that users preemptively
migrate away due to their rigid structure and incompatibility with other
environments.

This change affects the reviewdog module dependencies, but code changes
are limited to doghouse. The appengine runtime has been increased to
go112 for extended support and to enforce that appengine packages break
if used.

appengine/urlfetch.Client has been migrated to net/http.Client.
appengine/datastore has been migrated to cloud.google.com/go/datastore.
This took more work than the Client changes, since the two packages are
incompatible. There may be work to further refactor the abstractions,
but that is out of scope for this change.

appengine.IsDevServer has no clear successor, thus its usage and
conditional logic has been purged. If there is intereset to have a
comprehensive local development version of doghouse, this can be added
back, but you will lose all the other mocks that dev_appserver.py
offers: user auth, datastore access, etc.

appengine/log contains leveled print functions, Errorf/Infof/etc., that
have no clear parallel in the stdlib log package. As a stopgap,
[SEVERITY] prefixes have been added to all the new log.Printf calls.
This is not indended as a comprehensive solution, and it may be better
to move to another package for logging, say logrus or log15, but it is
out of scope at the moment.

Finally, a note: with the change from using appengine.Context to using
http.Request.Context, there are places where an http.Request and
context.Context are passed. This could be simplified, but is out of
scope.

Fixes #226