fix: GitLab CI MergeTrains
GitLab has a Feature called Merge Trains(see https://docs.gitlab.com/ee/ci/merge_request_pipelines/pipelines_for_merged_results/merge_trains/), which creates a "speculative" merge commit of the branch with the merge target and runs CI on that commit. This commit is not part of the Merge Request, therefore the logic to find the MR for the SHA that works without Merge Trains does not work any more.
Thankfully the env var "CI_MERGE_REQUEST_IID" is provided to get the MR for the pipeline.