fix: support determining build info on more GitHub Actions events
There are several GitHub Actions events that don't include a commit SHA
in the event payload fields that are currently parsed by Reviewdog. For
example, the `release` event only includes a `target_commitish` field
that does not contain a commit hash. This renders Reviewdog less
suitable for use in workflows triggered by such events.

To improve on the situation, let Reviewdog parse the `GITHUB_SHA`
environment variable to fetch a commit SHA if the already available
methods fail. As explained in the GitHub docs, this environment variable
is a useful last resort for other event types.