strict check of pr-review write permission
There is currently no way to check the available scope of the tokens provided by GitHub; the tokens available for GitHub Actions vary in permissions in complex ways, depending on the repository, workflow settings, and how the workflow is launched.

The most reliable way is to actually post review comments.