service/github: Fetch diff from base pull request repo
Previously, if you tried to fallback to Git based fetching with a private repo fork
on GitHub Actions, the command would fail because the automatic token used for
GitHub Actions is only scoped to the current repo, not the head repo.

GitHub forks store all commit data in the same namespace, which allows you to access
commit content from PRs / forks from the base repo. This should allow the use of the
default GitHub token to fetch the head PR content.